src/main/webapp/app/entities/prizes/delete/prizes-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { PrizesService } from '../service/prizes.service';

import { PrizesDeleteDialogComponent } from './prizes-delete-dialog.component';

describe('Prizes Management Delete Component', () => {
  let comp: PrizesDeleteDialogComponent;
  let fixture: ComponentFixture<PrizesDeleteDialogComponent>;
  let service: PrizesService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, PrizesDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(PrizesDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(PrizesDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(PrizesService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
